refactor: extract get_active_device_identities to identity.py
Move the large get_active_device_identities method (~725 lines) from
main.py to identity.py IdentityOperations mixin, continuing the
coordinator package refactoring.

Changes:
- Add get_active_device_identities method to IdentityOperations
- Add required imports for decrypt_eik, identity helpers, TokenCache
- Update test monkeypatching to use coordinator_identity and
  coordinator_registry modules
- Add per-file ruff ignores for coordinator package modules

Progress: main.py reduced from 4,561 to 3,739 lines
