+++ custom_components/googlefindmy/coordinator/polling.py
@@ -34,10 +34,10 @@ import asyncio
 import inspect
 import logging
 import time
-from collections.abc import Mapping
+from collections.abc import Callable, Mapping
 from datetime import datetime
 from statistics import mean, stdev
-from typing import TYPE_CHECKING, Any, Callable
+from typing import TYPE_CHECKING, Any
 
 from homeassistant.core import Event
 from homeassistant.helpers.event import async_call_later
@@ -63,7 +63,7 @@ class PollingOperations:
     """
 
     def _set_api_status(
-        self: "GoogleFindMyCoordinator", status: str, *, reason: str | None = None
+        self: GoogleFindMyCoordinator, status: str, *, reason: str | None = None
     ) -> None:
         """Update the API polling status and notify listeners if it changed."""
         if status == self._api_status_state and reason == self._api_status_reason:
@@ -80,7 +80,7 @@ class PollingOperations:
             pass
 
     def _set_fcm_status(
-        self: "GoogleFindMyCoordinator", status: str, *, reason: str | None = None
+        self: GoogleFindMyCoordinator, status: str, *, reason: str | None = None
     ) -> None:
         """Update the push transport status while avoiding noisy churn."""
         if status == self._fcm_status_state and reason == self._fcm_status_reason:
@@ -96,7 +96,7 @@ class PollingOperations:
             pass
 
     @property
-    def api_status(self: "GoogleFindMyCoordinator") -> StatusSnapshot:
+    def api_status(self: GoogleFindMyCoordinator) -> StatusSnapshot:
         """Return a snapshot describing the current API polling health."""
         return StatusSnapshot(
             state=self._api_status_state,
@@ -105,7 +105,7 @@ class PollingOperations:
         )
 
     @property
-    def fcm_status(self: "GoogleFindMyCoordinator") -> StatusSnapshot:
+    def fcm_status(self: GoogleFindMyCoordinator) -> StatusSnapshot:
         """Return a snapshot describing the current push transport health."""
         return StatusSnapshot(
             state=self._fcm_status_state,
@@ -114,21 +114,21 @@ class PollingOperations:
         )
 
     @property
-    def is_fcm_connected(self: "GoogleFindMyCoordinator") -> bool:
+    def is_fcm_connected(self: GoogleFindMyCoordinator) -> bool:
         """Convenience boolean for entities relying on push transport availability."""
         return self._fcm_status_state == FcmStatus.CONNECTED
 
     @property
-    def consecutive_timeouts(self: "GoogleFindMyCoordinator") -> int:
+    def consecutive_timeouts(self: GoogleFindMyCoordinator) -> int:
         """Return the number of consecutive poll timeouts."""
         return self._consecutive_timeouts
 
     @property
-    def last_poll_result(self: "GoogleFindMyCoordinator") -> str | None:
+    def last_poll_result(self: GoogleFindMyCoordinator) -> str | None:
         """Return the last recorded poll result ("success"/"failed")."""
         return self._last_poll_result
 
-    def _is_on_hass_loop(self: "GoogleFindMyCoordinator") -> bool:
+    def _is_on_hass_loop(self: GoogleFindMyCoordinator) -> bool:
         """Return True if currently executing on the HA event loop thread."""
         loop = self.hass.loop
         try:
@@ -137,7 +137,7 @@ class PollingOperations:
             return False
 
     def _run_on_hass_loop(
-        self: "GoogleFindMyCoordinator",
+        self: GoogleFindMyCoordinator,
         func: Callable[..., None],
         *args: Any,
         **kwargs: Any,
@@ -152,7 +152,7 @@ class PollingOperations:
         self.hass.loop.call_soon_threadsafe(func, *args, **kwargs)
 
     def _dispatch_async_request_refresh(
-        self: "GoogleFindMyCoordinator", *, task_name: str, log_context: str
+        self: GoogleFindMyCoordinator, *, task_name: str, log_context: str
     ) -> None:
         """Invoke ``async_request_refresh`` safely regardless of its implementation."""
         fn = getattr(self, "async_request_refresh", None)
@@ -175,7 +175,7 @@ class PollingOperations:
             self._run_on_hass_loop(_invoke)
 
     def _schedule_short_retry(
-        self: "GoogleFindMyCoordinator", delay_s: float = 5.0
+        self: GoogleFindMyCoordinator, delay_s: float = 5.0
     ) -> None:
         """Schedule a short, coalesced refresh instead of shifting the poll baseline.
 
@@ -219,7 +219,7 @@ class PollingOperations:
         else:
             self._run_on_hass_loop(_do_schedule)
 
-    async def _handle_dr_event(self: "GoogleFindMyCoordinator", _event: Event) -> None:
+    async def _handle_dr_event(self: GoogleFindMyCoordinator, _event: Event) -> None:
         """Handle Device Registry changes by rebuilding poll targets (rare)."""
         self._reindex_poll_targets_from_device_registry()
         # After changes, request a refresh so the next tick uses the new target sets.
@@ -229,7 +229,7 @@ class PollingOperations:
         )
 
     def _compute_type_cooldown_seconds(
-        self: "GoogleFindMyCoordinator", report_hint: str | None
+        self: GoogleFindMyCoordinator, report_hint: str | None
     ) -> int:
         """Return a server-aware cooldown duration in seconds for a crowdsourced report type.
 
@@ -257,7 +257,7 @@ class PollingOperations:
         return max(base_cooldown, effective_poll)
 
     def _apply_report_type_cooldown(
-        self: "GoogleFindMyCoordinator", device_id: str, report_hint: str | None
+        self: GoogleFindMyCoordinator, device_id: str, report_hint: str | None
     ) -> None:
         """Apply a per-device **poll** cooldown based on the crowdsourced report type.
 
@@ -287,7 +287,7 @@ class PollingOperations:
 
     # -------------------- Public read-only state for diagnostics/UI --------------------
     @property
-    def is_polling(self: "GoogleFindMyCoordinator") -> bool:
+    def is_polling(self: GoogleFindMyCoordinator) -> bool:
         """Expose current polling state (public read-only API).
 
         Returns:
@@ -296,7 +296,7 @@ class PollingOperations:
         return self._is_polling
 
     def get_fcm_acquire_duration_seconds(
-        self: "GoogleFindMyCoordinator",
+        self: GoogleFindMyCoordinator,
     ) -> float | None:
         """Duration between 'setup_start_monotonic' and 'fcm_acquired_monotonic'."""
         from .helpers.metrics import get_duration as _get_duration_impl
@@ -307,13 +307,13 @@ class PollingOperations:
         )
 
     def get_last_poll_duration_seconds(
-        self: "GoogleFindMyCoordinator",
+        self: GoogleFindMyCoordinator,
     ) -> float | None:
         """Duration of the most recent sequential polling cycle (if recorded)."""
         return self._get_duration("last_poll_start_mono", "last_poll_end_mono")
 
     # -------------------- FCM readiness checks --------------------
-    def _is_fcm_ready_soft(self: "GoogleFindMyCoordinator") -> bool:
+    def _is_fcm_ready_soft(self: GoogleFindMyCoordinator) -> bool:
         """Return True if push transport appears ready (no awaits, no I/O).
 
         Priority order:
@@ -358,7 +358,7 @@ class PollingOperations:
         except Exception:
             return False
 
-    def _note_fcm_deferral(self: "GoogleFindMyCoordinator", now_mono: float) -> None:
+    def _note_fcm_deferral(self: GoogleFindMyCoordinator, now_mono: float) -> None:
         """Advance a quiet escalation timeline while FCM is not ready.
 
         FIX: Use less aggressive log levels to reduce log spam (#124).
@@ -399,7 +399,7 @@ class PollingOperations:
                 reason="Push transport not connected after prolonged wait",
             )
 
-    def _clear_fcm_deferral(self: "GoogleFindMyCoordinator") -> None:
+    def _clear_fcm_deferral(self: GoogleFindMyCoordinator) -> None:
         """Clear the escalation timeline once FCM becomes ready (log once)."""
         if self._fcm_defer_started_mono:
             _LOGGER.info("FCM/push is ready; resuming scheduled polling.")
@@ -408,7 +408,7 @@ class PollingOperations:
         self._set_fcm_status(FcmStatus.CONNECTED)
 
     # -------------------- Poll timing prediction --------------------
-    def _get_predicted_poll_time(self: "GoogleFindMyCoordinator") -> float | None:
+    def _get_predicted_poll_time(self: GoogleFindMyCoordinator) -> float | None:
         """Predict the earliest next update time based on device histories."""
 
         history_store = getattr(self, "_device_update_history", None)
@@ -438,7 +438,7 @@ class PollingOperations:
 
     # -------------------- Push transport error handling --------------------
     def _note_push_transport_problem(
-        self: "GoogleFindMyCoordinator", cooldown_s: int = 90
+        self: GoogleFindMyCoordinator, cooldown_s: int = 90
     ) -> None:
         """Enter a temporary cooldown after a push transport failure to avoid spamming.
 
@@ -455,7 +455,7 @@ class PollingOperations:
             reason=f"Push transport recovering from error (cooldown {cooldown_s}s)",
         )
 
-    def force_poll_due(self: "GoogleFindMyCoordinator") -> None:
+    def force_poll_due(self: GoogleFindMyCoordinator) -> None:
         """Force the next poll to be due immediately (no private access required externally)."""
         effective_interval = max(self.location_poll_interval, self.min_poll_interval)
         # Move the baseline back so that (now - _last_poll_mono) >= effective_interval
