+++ tests/test_coordinator_snapshot.py
@@ -15,6 +15,8 @@ from custom_components.googlefindmy.coordinator import (
     _as_ha_attributes,
     _sync_get_last_gps_from_history,
 )
+from custom_components.googlefindmy.coordinator import main as coordinator_main
+from custom_components.googlefindmy.coordinator import registry as coordinator_registry
 
 
 class _DummyState:
@@ -75,10 +77,8 @@ def test_snapshot_uses_entry_scoped_unique_id(monkeypatch: pytest.MonkeyPatch) -
         "entry-1:device-42",
         "device_tracker.googlefindmy_device_42",
     )
-    monkeypatch.setattr(
-        "custom_components.googlefindmy.coordinator.main.er.async_get",
-        lambda hass: entity_registry,
-    )
+    # Use object-based patching (er is imported in registry.py)
+    monkeypatch.setattr(coordinator_registry.er, "async_get", lambda hass: entity_registry)
 
     hass = SimpleNamespace(states=_DummyStates())
     hass.states.set(
@@ -120,10 +120,8 @@ def test_snapshot_preserves_recorded_last_seen(monkeypatch: pytest.MonkeyPatch) 
         "entry-1:device-42",
         "device_tracker.googlefindmy_device_42",
     )
-    monkeypatch.setattr(
-        "custom_components.googlefindmy.coordinator.main.er.async_get",
-        lambda hass: entity_registry,
-    )
+    # Use object-based patching (er is imported in registry.py)
+    monkeypatch.setattr(coordinator_registry.er, "async_get", lambda hass: entity_registry)
 
     hass = SimpleNamespace(states=_DummyStates())
     iso_seen = "2024-02-03T12:34:56Z"
@@ -162,10 +160,8 @@ def test_snapshot_logs_formats_when_entity_missing(
     """Log should clarify which unique_id formats were considered when none match."""
 
     entity_registry = _DummyEntityRegistry()
-    monkeypatch.setattr(
-        "custom_components.googlefindmy.coordinator.main.er.async_get",
-        lambda hass: entity_registry,
-    )
+    # Use object-based patching (er is imported in registry.py)
+    monkeypatch.setattr(coordinator_registry.er, "async_get", lambda hass: entity_registry)
 
     hass = SimpleNamespace(states=_DummyStates())
