+++ tests/test_stats_sensor_updates.py
@@ -18,6 +18,8 @@ from custom_components.googlefindmy.const import (
     service_device_identifier,
 )
 from custom_components.googlefindmy.coordinator import GoogleFindMyCoordinator
+from custom_components.googlefindmy.coordinator import main as coordinator_main
+from custom_components.googlefindmy.coordinator import registry as coordinator_registry
 from custom_components.googlefindmy.sensor import (
     STATS_DESCRIPTIONS,
     GoogleFindMySemanticLabelSensor,
@@ -328,10 +330,8 @@ def test_history_fallback_increments_history_stat(
             f"{entry_id}:device-1",
             "device_tracker.googlefindmy_device_1",
         )
-        monkeypatch.setattr(
-            "custom_components.googlefindmy.coordinator.main.er.async_get",
-            lambda _hass: registry,
-        )
+        # Use object-based patching (er is imported in registry.py)
+        monkeypatch.setattr(coordinator_registry.er, "async_get", lambda _hass: registry)
 
         # No live state available -> force history fallback.
         hass.states = SimpleNamespace(get=lambda _eid: None)
